chore: move mobile check into globals store
